+++ CHANGELOG.md
@@ -167,6 +167,7 @@ Modifications can only be done via builders.
 - Make smart node more stable
 - Don't throw if builder is applied two times to the same scope
 - Optimize NodeBluePrint operator==
+- Optimize handling of realtime nodes
 
 ### Fixed